requirements.txt
fabric2
jinja2
pyyaml
kazoo